DELIA-63248 : [Auto][Sprint/Stable/Release] In video profile info 'codec' property is not available.
Reason for change: Inheritance Issue: The HlsStreamInfo structure inherits from StreamInfo, but it redefines the codecs. This can be confusing. Redefining fields in a derived class lead to unexpected behavior.

Test Procedure:
1) Play a HLS Stream via UVE Reference Player
2) Search for the following string, getAvailableVideoTracks
Risks: Low

DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_24_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_24_1
DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_24_1

+++ fragmentcollector_mpd.cpp
@@ -6781,7 +6781,7 @@ void StreamAbstractionAAMP_MPD::StreamSelection( bool newTune, bool forceSpeedsC
 	{
 		if(audioTrack.index == aTrackIdx)
 		{
-			mMediaStreamContext[eMEDIATYPE_AUDIO]->SetCurrentBandWidth(audioTrack.bandwidth);
+			mMediaStreamContext[eMEDIATYPE_AUDIO]->SetCurrentBandWidth((int)audioTrack.bandwidth);
 		}
 		bitratelist.push_back(audioTrack.bandwidth);
 	}
